requirements_dashboard.txt
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
xgboost==2.0.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
openai==1.3.8
anthropic==0.7.1
plotly==5.18.0
google-generativeai==0.3.0
